+++ apps/dashboard/src/lib/trpc/react.tsx
@@ -46,13 +46,28 @@ function TRPCReactProvider({
       links: [
         retryLink({
           retry: ({ error, attempts }) => {
-            if (error.data?.code === 'UNAUTHORIZED' && attempts < 2) {
-              void fetch(`${getDashboardUrl()}/api/auth/refresh-token`, {
+            if (
+              [
+                'FORBIDDEN',
+                'INTERNAL_SERVER_ERROR',
+                'NOT_FOUND',
+                'TOO_MANY_REQUESTS',
+              ].includes(error.data?.code ?? '') ||
+              error.message === 'Network request failed'
+            )
+              return false // Do not retry on specific errors
+
+            if (error.data?.code === 'UNAUTHORIZED') {
+              if (attempts > 1) return false // Do not retry more than once for unauthorized errors
+
+              fetch(`${getDashboardUrl()}/api/auth/refresh-token`, {
                 method: 'POST',
               })
-              return true
+
+              return true // Retry once after attempting to refresh the token
             }
-            return false
+
+            return attempts <= 3 // Retry up to 3 times for other errors
           },
           retryDelayMs: (attempts) => Math.min(1000 * 2 ** attempts, 30000),
         }),