+++ apps/mobile/src/lib/trpc.ts
@@ -1,3 +1,4 @@
+import type { QueryClient } from '@tanstack/react-query'
 import type { AppRouter } from '@yukinu/api'
 
 import { createTRPCClient, httpBatchLink, retryLink } from '@trpc/client'
@@ -5,29 +6,47 @@ import { createTRPCOptionsProxy } from '@trpc/tanstack-react-query'
 import { createQueryClient } from '@yukinu/lib/create-query-client'
 import SuperJSON from 'superjson'
 
-import { getAccessToken, getSessionToken, setAccessToken } from '@/lib/store'
+import { getAccessToken, getSessionToken } from '@/lib/secure-store'
 import { getBaseUrl } from '@/lib/utils'
 
-const queryClient = createQueryClient()
+let clientQueryClientSingleton: QueryClient | undefined
+const getQueryClient = () => {
+  if (typeof window === 'undefined') return createQueryClient()
+  return (clientQueryClientSingleton ??= createQueryClient())
+}
+
+const queryClient = getQueryClient()
 
 const trpcClient = createTRPCClient<AppRouter>({
   links: [
     retryLink({
       retry: ({ error, attempts }) => {
-        if (error.data?.code === 'UNAUTHORIZED' && attempts < 1) {
+        if (
+          [
+            'FORBIDDEN',
+            'INTERNAL_SERVER_ERROR',
+            'NOT_FOUND',
+            'TOO_MANY_REQUESTS',
+          ].includes(error.data?.code ?? '') ||
+          error.message === 'Network request failed'
+        )
+          return false // Do not retry on specific errors
+
+        if (error.data?.code === 'UNAUTHORIZED') {
+          if (attempts > 1) return false // Do not retry more than once for unauthorized errors
+
           fetch(`${getBaseUrl()}/api/auth/refresh-token`, {
             method: 'POST',
-            headers: { Authorization: `Bearer ${getSessionToken()}` },
+            headers: {
+              'Content-Type': 'application/json',
+              Authorization: `Bearer ${getSessionToken()}`,
+            },
           })
-            .then((res) => res.json() as Promise<{ accessToken: string }>)
-            .then(({ accessToken }) => setAccessToken(accessToken))
-            .catch(() => {
-              /* Ignore errors */
-            })
 
-          return true
+          return true // Retry once after attempting to refresh the token
         }
-        return false
+
+        return attempts <= 3 // Retry up to 3 times for other errors
       },
       retryDelayMs: (attempts) => Math.min(1000 * 2 ** attempts, 30000),
     }),
@@ -37,8 +56,8 @@ const trpcClient = createTRPCClient<AppRouter>({
       headers() {
         const headers = new Map<string, string>([['x-trpc-source', 'mobile']])
 
-        const accessToken = getAccessToken()
-        if (accessToken) headers.set('Authorization', `Bearer ${accessToken}`)
+        if (getAccessToken())
+          headers.set('Authorization', `Bearer ${getAccessToken()}`)
 
         return headers
       },