+++ apps/mobile/src/screens/login.tsx
@@ -1,97 +1,128 @@
-import type { AuthValidators } from '@yukinu/validators/auth'
+import type { GestureResponderEvent } from 'react-native'
 
 import { useNavigation } from '@react-navigation/native'
 import { useMutation } from '@tanstack/react-query'
-import { buttonVariants } from '@yukinu/ui/button'
-import * as React from 'react'
-import { Alert, Pressable, Text, TextInput, View } from 'react-native'
+import { AuthValidators } from '@yukinu/validators/auth'
+import { useState } from 'react'
+import { Alert, Linking, View } from 'react-native'
 
-import { setAccessToken, setSessionToken } from '@/lib/store'
+import { Button } from '@/components/ui/button'
+import { Input } from '@/components/ui/input'
+import { Text } from '@/components/ui/text'
+import { setAccessToken, setSessionToken } from '@/lib/secure-store'
+import { trpc } from '@/lib/trpc'
 import { getBaseUrl } from '@/lib/utils'
 
-export function LoginScreen() {
-  const navigation = useNavigation()
-  const [data, setData] = React.useState({ identifier: '', password: '' })
+type LoginInputKeys = keyof AuthValidators.LoginInput
 
-  const { mutate, isPending } = useMutation({
-    mutationKey: ['auth', 'login'],
-    mutationFn: async () => {
+export default function LoginScreen() {
+  const navigation = useNavigation()
+
+  const [data, setData] = useState<AuthValidators.LoginInput>({
+    identifier: '',
+    password: '',
+  })
+
+  const { mutate, error, isPending } = useMutation<
+    AuthValidators.LoginOutput,
+    Record<LoginInputKeys, string> & { message?: string },
+    GestureResponderEvent
+  >({
+    mutationFn: async (_: GestureResponderEvent) => {
+      const parsed = AuthValidators.loginInput.safeParse(data)
+      if (!parsed.success)
+        throw parsed.error.issues.reduce(
+          (acc, issue) => {
+            acc[issue.path[0] as LoginInputKeys] = issue.message
+            return acc
+          },
+          {} as Partial<Record<LoginInputKeys, string>>,
+        )
+
       const response = await fetch(`${getBaseUrl()}/api/auth/sign-in`, {
         method: 'POST',
         headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(data),
+        body: JSON.stringify(parsed.data),
       })
 
       if (!response.ok) throw new Error(await response.text())
+
       return response.json() as Promise<AuthValidators.LoginOutput>
     },
-    onSuccess: (data) => {
-      setSessionToken(data.token)
-      setAccessToken(data.accessToken)
-      navigation.navigate('main')
+    meta: { filter: trpc.user.profile.queryFilter() },
+    onSuccess: async (data) => {
+      await setSessionToken(data.token)
+      await setAccessToken(data.accessToken)
+      navigation.navigate('tabs', { screen: 'index', pop: true })
     },
-    onError: (error) => Alert.alert('Error', error.message),
+    onError: (error) =>
+      error.message && Alert.alert('Login Failed', error.message),
   })
 
   return (
-    <View className='bg-background flex-1'>
-      <View className='container h-full py-4 flex flex-col gap-4 justify-center items-center'>
-        <View className='w-full'>
-          <Text
-            className='text-2xl scroll-m-20 text-balance text-foreground'
-            style={{ fontFamily: 'Geist_600SemiBold' }}
-          >
+    <View className='flex-1 bg-background items-center justify-center px-4'>
+      <View className='w-full max-w-2xl gap-4 overflow-hidden rounded-xl bg-card py-4 ring-1 ring-foreground/10'>
+        <View className='grid auto-rows-min items-start gap-1 rounded-t-xl px-4'>
+          <Text className='text-center' variant='h4'>
             Login
           </Text>
-          <Text className='text-muted-foreground'>
-            Please enter your credentials to log in.
+          <Text className='text-muted-foreground text-sm'>
+            Welcome back! Please enter your credentials to log in.
           </Text>
         </View>
 
-        <View className='w-full flex flex-col gap-2'>
-          <Text
-            className='text-foreground'
-            style={{ fontFamily: 'Geist_500Medium' }}
-          >
-            Identifier
-          </Text>
-          <TextInput
-            placeholder='Enter your username or email'
-            className='text-foreground bg-input rounded-md px-2 placeholder:text-muted-foreground'
-            onChangeText={(text) => setData({ ...data, identifier: text })}
-            aria-disabled={isPending}
-            editable={!isPending}
-          />
-        </View>
-        <View className='w-full flex flex-col gap-2'>
-          <Text
-            className='text-foreground'
-            style={{ fontFamily: 'Geist_500Medium' }}
-          >
-            Password
-          </Text>
-          <TextInput
-            placeholder='Password'
-            className='text-foreground bg-input rounded-md px-2 placeholder:text-muted-foreground'
-            onChangeText={(text) => setData({ ...data, password: text })}
-            aria-disabled={isPending}
-            editable={!isPending}
-            secureTextEntry
-          />
+        <View className='px-4 gap-4'>
+          <View className='gap-2'>
+            <Text className='text-sm font-medium'>Email or Username</Text>
+            <Input
+              placeholder='Enter your identifier'
+              textContentType='username'
+              value={data.identifier}
+              onChangeText={(text) =>
+                setData((prev) => ({ ...prev, identifier: text }))
+              }
+              editable={!isPending}
+            />
+            {error?.identifier && (
+              <Text className='text-destructive text-sm'>
+                {error.identifier}
+              </Text>
+            )}
+          </View>
+
+          <View className='gap-2'>
+            <Text className='text-sm font-medium'>Password</Text>
+            <Input
+              placeholder='Enter your password'
+              textContentType='password'
+              value={data.password}
+              onChangeText={(text) =>
+                setData((prev) => ({ ...prev, password: text }))
+              }
+              editable={!isPending}
+              secureTextEntry
+            />
+            {error?.password && (
+              <Text className='text-destructive text-sm'>{error.password}</Text>
+            )}
+          </View>
         </View>
 
-        <Pressable
-          className={buttonVariants({ className: 'w-full' })}
-          onPress={() => mutate()}
-          disabled={isPending}
-        >
-          <Text
-            className='text-primary-foreground'
-            style={{ fontFamily: 'Geist_500Medium' }}
-          >
-            Login
+        <View className='items-center rounded-b-xl px-4'>
+          <Button className='w-full' onPress={mutate} disabled={isPending}>
+            <Text>Login</Text>
+          </Button>
+
+          <Text className='w-full mt-4 text-sm text-muted-foreground'>
+            Don&#39;t have an account?{' '}
+            <Text
+              className='text-sm text-primary'
+              onPress={() => Linking.openURL(`${getBaseUrl()}/register`)}
+            >
+              Register now
+            </Text>
           </Text>
-        </Pressable>
+        </View>
       </View>
     </View>
   )