+++ apps/mobile/src/screens/not-found.tsx
@@ -1,31 +1,24 @@
 import { useNavigation } from '@react-navigation/native'
-import { buttonVariants } from '@yukinu/ui/button'
-import { Pressable, Text, View } from 'react-native'
+import { View } from 'react-native'
 
-export function NotFoundScreen() {
+import { Button } from '@/components/ui/button'
+import { Text } from '@/components/ui/text'
+
+export default function NotFoundScreen() {
   const navigation = useNavigation()
 
   return (
-    <View className='bg-background py-4 flex-1'>
-      <View className='container h-full flex flex-col items-center justify-center gap-6'>
-        <Text
-          className='text-4xl scroll-m-20 text-balance text-foreground'
-          style={{ fontFamily: 'Geist_700Bold' }}
-        >
-          404 - Not Found
-        </Text>
+    <View className='flex-1 bg-background'>
+      <View className='container flex-1 items-center justify-center gap-6'>
+        <Text variant='h1'>404 - Not Found</Text>
 
-        <Pressable
-          className={buttonVariants({ size: 'lg' })}
-          onPress={() => navigation.navigate('main', undefined, { pop: true })}
+        <Button
+          onPress={() =>
+            navigation.navigate('tabs', { screen: 'index', pop: true })
+          }
         >
-          <Text
-            className='text-primary-foreground'
-            style={{ fontFamily: 'Geist_500Medium' }}
-          >
-            Take me home
-          </Text>
-        </Pressable>
+          <Text>Take me Home</Text>
+        </Button>
       </View>
     </View>
   )