+++ apps/web/app/(main)/[slug]/_components/product-images.tsx
@@ -18,12 +18,10 @@ export const ProductImages: React.FC = () => {
 
       <div className='relative aspect-square w-full rounded-md border'>
         <Image
-          src={currentImage ?? '/assets/logo.svg'}
+          src={currentImage ?? '/favicon.svg'}
           alt={`${id}-main-image`}
           sizes='(max-width: 768px) 100vw, (max-width: 1200px) 50vw, 33vw'
-          className={cn('rounded-md object-contain p-0.5', {
-            'dark:invert': !currentImage,
-          })}
+          className='rounded-md object-contain p-0.5'
           fill
         />
       </div>