+++ apps/mobile/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @yukinu/mobile
 
+## 0.1.0
+
+### Minor Changes
+
+- [#314](https://github.com/tiesen243/yukinu/pull/314) [`7cecd87`](https://github.com/tiesen243/yukinu/commit/7cecd87b3cad0b1f3d585ec3e258ea21dcc48ce9) Thanks [@tiesen243](https://github.com/tiesen243)! - update logo
+
+### Patch Changes
+
+- Updated dependencies [[`7cecd87`](https://github.com/tiesen243/yukinu/commit/7cecd87b3cad0b1f3d585ec3e258ea21dcc48ce9)]:
+  - @yukinu/api@0.8.0
+  - @yukinu/lib@0.6.0
+  - @yukinu/ui@0.7.0
+  - @yukinu/validators@0.8.0
+
+## 0.0.3
+
+### Patch Changes
+
+- [#312](https://github.com/tiesen243/yukinu/pull/312) [`1ef3a4a`](https://github.com/tiesen243/yukinu/commit/1ef3a4a16c35cc8cd69344a1aea1f788ea313676) Thanks [@tiesen243](https://github.com/tiesen243)! - update mobile app
+
+- [#312](https://github.com/tiesen243/yukinu/pull/312) [`4fc10aa`](https://github.com/tiesen243/yukinu/commit/4fc10aa6f5e8d78031dcb92cc64df2bb11863160) Thanks [@tiesen243](https://github.com/tiesen243)! - extract button components
+
+## 0.0.2
+
+### Patch Changes
+
+- [#310](https://github.com/tiesen243/yukinu/pull/310) [`661f387`](https://github.com/tiesen243/yukinu/commit/661f387784c868faee546b2a6c1a1c364b3bcf53) Thanks [@tiesen243](https://github.com/tiesen243)! - update retryLink config
+
+- [#310](https://github.com/tiesen243/yukinu/pull/310) [`3e3267a`](https://github.com/tiesen243/yukinu/commit/3e3267a85c62e3f11bf4955524a1df8f32eb2b49) Thanks [@tiesen243](https://github.com/tiesen243)! - added redux
+
 ## 0.0.1
 
 ### Patch Changes